Chapter 7 - Step defs from inside organize code
I have taken the Account, Teller, and CashSlot classes and created 3
java class files in src/main.  The book state 'Since Steps is also in
NiceBank pckg, no imports are needed'.  I understand the no imports but
isn't this in a different NiceBank pckg.  Src/main and not Src/test?  I
need answers!!!

Test runs though
