+++ 05. Step Defs on the outside/ATM/src/main/java/nicebank/Account.java
@@ -1,0 +1,13 @@
+package nicebank;
+
+public class Account {
+	private Money balance = new Money();
+	 
+    public void deposit(Money amount) {
+    	balance = balance.add(amount);
+    }
+    
+    public Money getBalance(){
+    	return balance;
+    }
+}
