+++ 05. Step Defs on the outside/ATM/src/main/java/nicebank/CashSlot.java
@@ -1,0 +1,15 @@
+package nicebank;
+
+public class CashSlot {
+	 private int contents;
+	  
+	  public int getContents(){
+		 return contents; 
+	  }
+	  
+	  public void dispense(int dollars){
+		  contents = dollars;
+	  }
+
+
+}
