+++ 05. Step Defs on the outside/ATM/src/main/java/nicebank/Teller.java
@@ -1,0 +1,13 @@
+package nicebank;
+
+public class Teller {
+	  private CashSlot cashSlot;
+	  
+	  public Teller (CashSlot cashSlot){
+		  this.cashSlot = cashSlot;
+	  }
+	  
+	  public void withdrawFrom(Account account, int dollars){
+		  cashSlot.dispense(dollars);
+	  }
+}
